requirements.txt
requests>=2.31,<3
pyyaml>=6.0,<7
numpy>=1.26,<2
pandas>=2.1,<3
ccxt>=4.0,<5
ta>=0.11,<1
tweepy>=4.14,<5
openai>=1.0,<2
anthropic>=0.39,<1
pydantic>=2.5,<3
python-dotenv>=1.0,<2
streamlit>=1.29,<2
plotly>=5.18,<6
